requirements.txt
# TSD HFT Engine - Dependencies
# Lean and intentional. Every dependency earns its place.

# --- Web Framework (ONE framework, async-native) ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
jinja2==3.1.4

# --- HTTP Client ---
httpx==0.27.2

# --- Configuration ---
pyyaml==6.0.2
python-dotenv==1.0.1

# --- Exchange Integration ---
websocket-client==1.8.0

# --- Data & Analysis ---
pandas==2.2.3
numpy>=2.1.0
ta==0.11.0

# --- Database (trade log, audit trail) ---
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.30.0
aiosqlite==0.20.0

# --- Async ---
aiofiles==24.1.0

# --- Security ---
python-jose==3.3.0
cryptography==43.0.0

# --- Monitoring ---
psutil==6.1.0

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# --- Logging ---
python-json-logger==2.0.7
